+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+- Handle multiple inline links ([#1])
+
 ### 1.0.0
 
 - Initial release.
+
+<!-- LINKS -->
+
+[#1]:https://github.com/godaddy/dmd/pull/1